ccxcxcx

+++ waste-manager-service/src/main/java/com/prueba/addressservice/repository/WasteManagerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WasteManagerRepository extends JpaRepository<WasteManager, Integer> {
+public interface WasteManagerRepository extends JpaRepository<WasteManager, Long> {
 }
